refactor: extract logic out of WebhookController
