+++ github_webhook_handler/src/server.ts
@@ -2,8 +2,8 @@ import dotenv from "dotenv";
 dotenv.config();
 import express from "express";
 import bodyParser from "body-parser";
-import { verifySignature } from './services/github'
 import { handleWebhook } from './controllers/webHookController'
+import { verifySignature } from './services/github/verifySignature'
 
 const app = express();
 app.use(bodyParser.json());
