+++ github_webhook_handler/src/services/openai/generateCodeReview.ts
@@ -1,7 +1,7 @@
 import { OpenAI } from "openai";
 import { OPENAI_API_KEY } from '../../config'
-import { postReviewComment } from '../github'
 import { GenerateCodeReviewParams } from './types'
+import { postReviewComment } from '../github/postReviewComment'
 
 const openai = new OpenAI({
   apiKey: OPENAI_API_KEY,
